# metric, deps: fork prometheus repo and improve histogram lookup perf