--- a/pkg/ccl/testccl/sqlstatsccl/sql_stats_test.go
+++ b/pkg/ccl/testccl/sqlstatsccl/sql_stats_test.go
@@ -84,26 +84,6 @@ func TestSQLStatsRegions(t *testing.T) {
 	}()
 
 	tdb := sqlutils.MakeSQLRunner(host.ServerConn(0))
-
-	// Max out rate limit for replica rebalances.
-	tdb.Exec(t, `set cluster setting kv.snapshot_rebalance.max_rate = '1g'`)
-
-	// Shorten the closed timestamp target duration so that span configs
-	// propagate more rapidly.
-	tdb.Exec(t, `SET CLUSTER SETTING kv.closed_timestamp.target_duration = '200ms'`)
-	tdb.Exec(t, "SET CLUSTER SETTING kv.allocator.load_based_rebalancing = off")
-
-	// Lengthen the lead time for the global tables to prevent overload from
-	// resulting in delays in propagating closed timestamps and, ultimately
-	// forcing requests from being redirected to the leaseholder. Without this
-	// change, the test sometimes is flakey because the latency budget allocated
-	// to closed timestamp propagation proves to be insufficient. This value is
-	// very cautious, and makes this already slow test even slower.
-	tdb.Exec(t, "SET CLUSTER SETTING kv.closed_timestamp.side_transport_interval = '10ms'")
-	tdb.Exec(t, `SET CLUSTER SETTING kv.closed_timestamp.lead_for_global_reads_override = '500ms'`)
-	tdb.Exec(t, "SET CLUSTER SETTING kv.rangefeed.closed_timestamp_refresh_interval = '10ms'")
-	tdb.Exec(t, `ALTER TENANT ALL SET CLUSTER SETTING spanconfig.reconciliation_job.checkpoint_interval = '500ms'`)
-
 	tdb.Exec(t, `ALTER RANGE meta configure zone using constraints = '{"+region=gcp-us-west1": 1, "+region=gcp-us-central1": 1, "+region=gcp-us-east1": 1}';`)
 
 	// Create secondary tenants