--- a/pkg/util/metric/metric_test.go
+++ b/pkg/util/metric/metric_test.go
@@ -17,6 +17,7 @@ import (
 	"time"
 
 	_ "github.com/cockroachdb/cockroach/pkg/util/log" // for flags
+	"github.com/cockroachdb/cockroach/pkg/util/randutil"
 	"github.com/kr/pretty"
 	"github.com/prometheus/client_golang/prometheus"
 	prometheusgo "github.com/prometheus/client_model/go"
@@ -934,5 +935,36 @@ func TestHistogramVec(t *testing.T) {
 
 		require.Equal(t, uint64(1), *metrics[1].Histogram.SampleCount)
 		require.Equal(t, float64(1), *metrics[1].Histogram.SampleSum)
+	})
+}
+
+func BenchmarkHistogramRecordValue(b *testing.B) {
+	h := NewHistogram(HistogramOptions{
+		Metadata: Metadata{
+			Name:       "my.test.metric",
+			MetricType: prometheusgo.MetricType_HISTOGRAM,
+		},
+		Duration:     0,
+		BucketConfig: IOLatencyBuckets,
+		Mode:         HistogramModePrometheus,
+	})
+
+	b.ResetTimer()
+	r, _ := randutil.NewTestRand()
+
+	b.Run("insert integers", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			h.RecordValue(int64(i))
+		}
+	})
+	b.Run("insert zero", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			h.RecordValue(0)
+		}
+	})
+	b.Run("random integers", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			h.RecordValue(int64(randutil.RandIntInRange(r, int(IOLatencyBuckets.min), int(IOLatencyBuckets.max))))
+		}
 	})
 }