… build/bazelutil/distdir_files.bzl
… go.sum